chapter4/execution-tools/requirements.txt
# MCP and Core Dependencies
mcp>=0.9.0
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=1.0.0

# External API Integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
PyGithub>=2.0.0

# Process Management
subprocess-tee>=0.4.0
psutil>=5.9.8

# Scientific Computing - Core
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Machine Learning & AI
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0

# Deep Learning (Optional, comment out if not needed)
# torch>=2.1.0
# tensorflow>=2.14.0

# Data Processing & Analysis
openpyxl>=3.1.0
xlrd>=2.0.0
XlsxWriter>=3.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Scientific Libraries
statsmodels>=0.14.0
scikit-image>=0.22.0
pillow>=10.2.0
opencv-python>=4.9.0
plotly>=5.18.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
PyMuPDF>=1.24.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.12.0

# Utilities
requests>=2.31.0
pydantic>=2.6.0
structlog>=23.1.0
tabulate>=0.9.0
colorama>=0.4.4
